used SimpleCallback instead of Callback - shorter version
