--- app/src/main/java/com/example/myshoppinglist/presentation/helper/ItemTouchHelperAdapter.kt
@@ -1,6 +1,0 @@
-package com.example.myshoppinglist.presentation.helper
-
-interface ItemTouchHelperAdapter {
-    fun onItemMove(fromPosition: Int, toPosition: Int)
-    fun onItemDismiss(position: Int)
-}
