--- app/src/main/java/com/example/myshoppinglist/presentation/helper/ItemTouchHelperCallback.kt
@@ -1,35 +1,0 @@
-package com.example.myshoppinglist.presentation.helper
-
-import androidx.recyclerview.widget.ItemTouchHelper
-import androidx.recyclerview.widget.RecyclerView
-
-class ItemTouchHelperCallback(private val mAdapter: ItemTouchHelperAdapter) :
-    ItemTouchHelper.Callback() {
-
-    override fun getMovementFlags(
-        recyclerView: RecyclerView,
-        viewHolder: RecyclerView.ViewHolder
-    ): Int {
-        val dragFlags = ItemTouchHelper.UP or ItemTouchHelper.DOWN
-        val swipeFlags = ItemTouchHelper.START or ItemTouchHelper.END
-        return makeMovementFlags(dragFlags, swipeFlags)
-    }
-
-    override fun onMove(
-        recyclerView: RecyclerView,
-        viewHolder: RecyclerView.ViewHolder,
-        target: RecyclerView.ViewHolder
-    ): Boolean {
-//        mAdapter.onItemMove(
-//            viewHolder.adapterPosition,
-//            target.adapterPosition
-//        );
-// TODO("Переопределить для изменения положения элементов в списке")
-        return false;
-    }
-
-
-    override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-        mAdapter.onItemDismiss(viewHolder.adapterPosition)
-    }
-}
