--- app/src/main/java/com/example/myshoppinglist/presentation/helper/ItemTouchHelperViewHolder.kt
@@ -1,6 +1,0 @@
-package com.example.myshoppinglist.presentation.helper
-
-interface ItemTouchHelperViewHolder {
-    fun onItemSelected()
-    fun onItemClear()
-}
